Refactor the chaos engine controller package
